将音频封装成Plugin以DLL的形式加载，通过CMakeLists开关是否开启音频功能，默认为不加载音频插件。
