+++ Tests/Audio/TestAudio.cpp
@@ -1,4 +1,5 @@
-#include <Audio/AudioEngine.hpp>
+#ifdef ENABLE_AUDIO
+#include <AudioEngine.hpp>
 #include <thread>
 
 bool UnitTestAudio() {
@@ -12,8 +13,18 @@ bool UnitTestAudio() {
 		Engine.GetManager()->PlaySound("../Assets/Media/Audio/sample1.wav");
 
 		// µÈ´ý²¥·Å
-		std::this_thread::sleep_for(std::chrono::seconds(60));
+		std::this_thread::sleep_for(std::chrono::seconds(10));
 		Engine.Shutdown();
 	}
 	return true;
 }
+
+#else 
+#include <Core/EngineLogger.hpp>
+
+bool UnitTestAudio() {
+	LOG_INFO("Not enable plugins audio. Skip audio unit test. If you want to do audio test, set ENABLE_PLUGINS_AUDIO ON on CMakeLists.");
+	return true;
+}
+
+#endif
